+++ Metal.NET/Metal/MTLAccelerationStructurePassSampleBufferAttachmentDescriptor.cs
@@ -17,7 +17,7 @@ public class MTLAccelerationStructurePassSampleBufferAttachmentDescriptor : IDis
     public nuint EndOfEncoderSampleIndex
     {
         get => ObjectiveCRuntime.MsgSendNUInt(NativePtr, MTLAccelerationStructurePassSampleBufferAttachmentDescriptorSelector.EndOfEncoderSampleIndex);
-        set => ObjectiveCRuntime.MsgSend(NativePtr, MTLAccelerationStructurePassSampleBufferAttachmentDescriptorSelector.SetEndOfEncoderSampleIndex, (nuint)value);
+        set => ObjectiveCRuntime.MsgSend(NativePtr, MTLAccelerationStructurePassSampleBufferAttachmentDescriptorSelector.SetEndOfEncoderSampleIndex, value);
     }
 
     public MTLCounterSampleBuffer SampleBuffer
@@ -29,7 +29,7 @@ public class MTLAccelerationStructurePassSampleBufferAttachmentDescriptor : IDis
     public nuint StartOfEncoderSampleIndex
     {
         get => ObjectiveCRuntime.MsgSendNUInt(NativePtr, MTLAccelerationStructurePassSampleBufferAttachmentDescriptorSelector.StartOfEncoderSampleIndex);
-        set => ObjectiveCRuntime.MsgSend(NativePtr, MTLAccelerationStructurePassSampleBufferAttachmentDescriptorSelector.SetStartOfEncoderSampleIndex, (nuint)value);
+        set => ObjectiveCRuntime.MsgSend(NativePtr, MTLAccelerationStructurePassSampleBufferAttachmentDescriptorSelector.SetStartOfEncoderSampleIndex, value);
     }
 
     public static implicit operator nint(MTLAccelerationStructurePassSampleBufferAttachmentDescriptor value)
